pyproject.toml
[tool.poetry]
name = "data-dictionary-cui-mapping"
version = "1.0.5"
description = "This package allows you to load in a data dictionary and map cuis to defined fields using either the UMLS API or MetaMap API from NLM."
authors = ["Kevin Armengol <kevin.armengol@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/kevon217/data-dictionary-cui-mapping"
keywords = ["BRICS", "Data Dictionary", "UMLS", "MetaMap", "CUI", "NLM"]
include = ["data_dictionary_cui_mapping/**/*"]

[tool.poetry.dependencies]
python = "^3.8.1"
pandas = ">=1.5.2"
openpyxl = ">=3.0.10"
requests = ">=2.28.1"
python-dotenv = "0.21.1"
requests-html = ">=0.10.0"
hydra-core = ">=1.1.0"
omegaconf = "2.1.2"
prefect = {extras = ["viz"], version = "^2.8.3"}
transformers = "4.18.0"
pathlib = "^1.0.1"
pinecone-client = "^2.2.1"
pandas-log = "^0.1.7"
sentence-transformers = "^2.2.2"
splade = {git = "https://github.com/naver/splade.git"}
cchardet = "^2.1.7"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
radon = "^5.1.0"
vulture = "^2.7"
flake8 = "^6.0.0"
pre-commit = "^3.0.4"
isort = "^5.11.3"
jupyterlab = "^3.6.1"
jupyter = "^1.0.0"
wheel = "^0.38.4"
deptry = "^0.8.0"
bs4 = "^0.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
